avniproject/avni-server#813 | Fix test failure

+++ avni-server-api/src/main/java/org/avni/server/dao/SubjectSearchRepository.java
@@ -15,6 +15,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.StringUtils;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -52,7 +53,7 @@ public class SubjectSearchRepository extends RoleSwitchableRepository {
     }
 
     @Transactional
-    public Long getTotalCount(SubjectSearchRequest searchRequest, SearchBuilder searchBuilder) {
+    public BigInteger getTotalCount(SubjectSearchRequest searchRequest, SearchBuilder searchBuilder) {
         SubjectType subjectType = StringUtils.isEmpty(searchRequest.getSubjectType()) ? null : subjectTypeRepository.findByUuid(searchRequest.getSubjectType());
         SqlQuery query = searchBuilder.getSQLCountQuery(searchRequest, subjectType);
         try {
@@ -62,7 +63,7 @@ public class SubjectSearchRepository extends RoleSwitchableRepository {
                 sql.setParameter(name, value);
             });
 
-            return (Long) sql.getSingleResult();
+            return (BigInteger) sql.getSingleResult();
         } finally {
             setRoleBackToUser();
         }
